requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
pillow
drf-yasg
python-decouple
gunicorn
whitenoise
dj-database-url
django-rest-passwordreset